marimo_note/pyproject.toml
[project]
name = "python"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10.5"
dependencies = [
    "duckdb>=1.3.2",
    "marimo>=0.14.10",
    "pip>=25.1.1",
    "polars>=1.31.0",
    "psycopg2-binary>=2.9.10",
    "pyarrow>=20.0.0",
    "sqlalchemy>=2.0.41",
    "sqlglot>=27.0.0",
]

[tool.uv.workspace]
members = [
    "api_server_templates/litestar_template",
    "project_templates/django_template",
]
